requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx>=0.27.0
python-dotenv>=1.0.0
google-genai>=1.0.0
slowapi>=0.1.9
